+++ src/main/java/org/codinjutsu/tools/jenkins/view/action/LoadBuildsAction.java
@@ -3,10 +3,10 @@ package org.codinjutsu.tools.jenkins.view.action;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.UpdateInBackground;
-import com.intellij.openapi.progress.ProgressIndicator;
-import com.intellij.openapi.progress.Task;
 import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.project.Project;
+import org.codinjutsu.tools.jenkins.logic.JenkinsBackgroundTask;
+import org.codinjutsu.tools.jenkins.logic.JenkinsBackgroundTaskFactory;
 import org.codinjutsu.tools.jenkins.logic.RequestManagerInterface;
 import org.codinjutsu.tools.jenkins.model.Job;
 import org.codinjutsu.tools.jenkins.model.JobType;
@@ -30,35 +30,29 @@ public class LoadBuildsAction extends AnAction implements DumbAware, UpdateInBac
     }
 
     private void loadBuildsForSelectedJob(Project project) {
-        final BrowserPanel browserPanel = BrowserPanel.getInstance(project);
-        Optional.ofNullable(browserPanel.getSelectedJob()).ifPresent(job -> loadBuilds(project, job));
+        Optional.ofNullable(BrowserPanel.getInstance(project).getSelectedJob()).ifPresent(job -> loadBuilds(project, job));
     }
 
     public void loadBuilds(Project project, Job job) {
-        final BrowserPanel browserPanel = BrowserPanel.getInstance(project);
         final boolean expandAfterLoad = job.getLastBuilds().isEmpty();
-        try {
-            new Task.Backgroundable(project, getTemplatePresentation().getText(), false) {
+        JenkinsBackgroundTaskFactory.getInstance(project).createBackgroundTask(getTemplatePresentation().getText(),
+                new JenkinsBackgroundTask.JenkinsTask() {
 
-                @Override
-                public void onSuccess() {
-                    super.onSuccess();
-                    browserPanel.refreshJob(job);
-                    if (expandAfterLoad) {
-                        browserPanel.expandSelectedJob();
+                    @Override
+                    public void onSuccess() {
+                        JenkinsBackgroundTask.JenkinsTask.super.onSuccess();
+                        final BrowserPanel browserPanel = BrowserPanel.getInstance(project);
+                        browserPanel.refreshJob(job);
+                        if (expandAfterLoad) {
+                            browserPanel.expandSelectedJob();
+                        }
                     }
-                }
 
-                @Override
-                public void run(@NotNull ProgressIndicator progressIndicator) {
-                    progressIndicator.setIndeterminate(true);
-                    RequestManagerInterface requestManager = browserPanel.getJenkinsManager();
-                    job.setLastBuilds(requestManager.loadBuilds(job));
-                }
-            }.queue();
-        } catch (Exception ex) {
-            browserPanel.notifyErrorJenkinsToolWindow("Unable to load builds: " + ex.getMessage());
-        }
+                    @Override
+                    public void run(@NotNull RequestManagerInterface requestManager) {
+                        job.setLastBuilds(requestManager.loadBuilds(job));
+                    }
+                }).queue();
     }
 
     @Override